fix(lprnet):fix bugs for 3.0.0 1.fix README bugs 2.change sail.IOMode.SYSO in lprnet_sail_bmcv_sail.py 3.remove wrong log for python scripy 4.fix warning msg for cpp compilation 5.download lmdb to generate int8bmodel
(cherry picked from commit 8034922ab030f742443db7e59866aac07103af2a)

+++ simple/lprnet/cpp/lprnet_cv_cv_bmrt/main.cpp
@@ -131,7 +131,7 @@ int main(int argc, char** argv) {
         string img_name = img_file.substr(index + 1);
         batch_imgs.push_back(img);
         batch_names.push_back(img_name);
-        if (batch_imgs.size() == batch_size) {
+        if ((int)batch_imgs.size() == batch_size) {
           vector<string> results = detect(net, batch_imgs, ts);
           for(int i = 0; i < batch_size; i++){
             fs::path image_file(batch_names[i]);
